requirements.txt
aniso8601==3.0.2
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.5.1
coveralls==1.3.0
docopt==0.6.2
Flask==1.0.2
flask-restplus==0.11.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
more-itertools==4.2.0
pluggy==0.6.0
py==1.5.3
PyJWT==1.6.4
pytest==3.6.1
pytest-cov==2.5.1
pytz==2018.4
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1